+++ packages/rich-domain-cli/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.5.1
+
+### Patch Changes
+
+- some fixes
+
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
 ### [0.4.2](https://github.com/tarcisioandrade/rich-domain/compare/v1.8.4...v0.4.2) (2026-01-25)